Require azure/core manually
'azure' moved part of it to a gem called 'azure_core'. They released a
version that lists that gem as a dependency. When this version is
used with 'fog-azure', we get an error because 'azure/core' isn't loaded
(see https://github.com/Azure/azure-sdk-for-ruby/pull/383).

We can simply preload it in fog-azure and not let the error in the
dependency affect us.